requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
python-dotenv>=1.0.0
stripe>=7.6.0
httpx>=0.25.1
pydantic>=2.5.2
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
google-generativeai>=0.3.1
email-validator>=2.1.0